feat: add user profile, matching, and enhanced profile view with location utilities
